refactor(profile/sections): reexport files in `index.ts` file
