+++ app/profile/recently-played/page.tsx
@@ -1,4 +1,4 @@
-import { ProfileRecentlyPlayedSection } from '../sections/recently-played'
+import { ProfileRecentlyPlayedSection } from '../sections'
 
 export default async function ProfileRecentlyPlayedPage() {
   return <ProfileRecentlyPlayedSection limit={50} />
