+++ app/profile/sections/index.ts
@@ -1,0 +1,6 @@
+export * from './items'
+export * from './top-items'
+export * from './top-artists'
+export * from './top-tracks'
+export * from './top-genres'
+export * from './recently-played'
