+++ app/profile/top/tracks/page.tsx
@@ -3,7 +3,7 @@ import { cookies } from 'next/headers'
 import { ACCESS_TOKEN } from '@app/api/constants'
 import { getTopTracks } from '@app/api/fetchers'
 import { PageProps } from '@app/types'
-import { TopItemsSection } from '@app/profile/sections/top-items'
+import { TopItemsSection } from '@app/profile/sections'
 import { validateTimeRange } from '@app/utils/time-range'
 import { validateView } from '@app/utils/view'
 import { TIME_RANGE, VIEW } from '@app/constants'
